api/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Utilities
python-json-logger==2.0.7
pydantic==2.5.0

# Optional: For better performance
# onnxruntime-gpu==1.16.3  # Uncomment for GPU support